requirements.txt
dnspython==2.2.1
Faker==15.1.1
kafka-python==2.0.2
pymongo==4.3.2
python-dateutil==2.8.2
six==1.16.0